Split out additional settings component
